+++ src/surfaces/base/advanced-surface.js
@@ -36,22 +36,6 @@ class AdvancedSurface extends Surface {
   }
 
   /**
-   * Sets a custom identifier that must be unique for all views on a per-team basis
-   *
-   * **Slack Validation Rules:**
-   *    * Max 255 characters
-   *
-   * {@link https://api.slack.com/reference/surfaces/views|View in Slack API Documentation}
-   *
-   * @param {string} string
-   * @return {this} The instance on which the method is called
-   */
-
-  externalId(string) {
-    return this.set(string, props.externalId);
-  }
-
-  /**
    * When called, builds the view and prints to the console the preview URL in
    * order to open and preview the view on the Slack Block Builder website
    */
@@ -59,6 +43,7 @@ class AdvancedSurface extends Surface {
   printPreviewUrl() {
     this.build();
 
+    // eslint-disable-next-line no-console
     console.log(encodeURI(`https://app.slack.com/block-kit-builder/#${JSON.stringify(this.result)}`).replace(/[!'()*]/g, escape));
   }
 }